feat: change to permission framework

+++ plugins/qeta-backend/src/service/routes/answers.ts
@@ -1,10 +1,9 @@
 import Ajv from 'ajv';
 import addFormats from 'ajv-formats';
 import {
-  checkPermissions,
+  authorize,
   getCreated,
   getUsername,
-  isModerator,
   mapAdditionalFields,
 } from '../util';
 import {
@@ -16,7 +15,13 @@ import {
 import { Request, Router } from 'express';
 import {
   qetaCreateAnswerPermission,
-  qetaReadPermission,
+  qetaCreateCommentPermission,
+  qetaDeleteAnswerPermission,
+  qetaDeleteCommentPermission,
+  qetaEditAnswerPermission,
+  qetaEditQuestionPermission,
+  qetaReadAnswerPermission,
+  qetaReadQuestionPermission,
 } from '@drodil/backstage-plugin-qeta-common';
 import { Response } from 'express-serve-static-core';
 import {
@@ -34,7 +39,6 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
   router.get(`/answers`, async (request, response) => {
     // Validation
     const username = await getUsername(request, options, true);
-    await checkPermissions(request, qetaReadPermission, options);
     const validateQuery = ajv.compile(AnswersQuerySchema);
     if (!validateQuery(request.query)) {
       response
@@ -62,7 +66,6 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
   // POST /questions/:id/answers
   router.post(`/questions/:id/answers`, async (request, response) => {
     // Validation
-    await checkPermissions(request, qetaCreateAnswerPermission, options);
     const validateRequestBody = ajv.compile(PostAnswerSchema);
     if (!validateRequestBody(request.body)) {
       response
@@ -79,8 +82,16 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
       return;
     }
 
+    await authorize(
+      request,
+      qetaReadQuestionPermission,
+      options,
+      request.params.id,
+    );
+
+    await authorize(request, qetaCreateAnswerPermission, options);
+
     const username = await getUsername(request, options);
-    const moderator = await isModerator(request, options);
     const created = await getCreated(request, options);
     const question = await database.getQuestion(username, questionId, false);
     if (!question) {
@@ -107,7 +118,7 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
 
     const followingUsers = await database.getUsersForTags(question.tags);
     notificationMgr.onNewAnswer(username, question, answer, followingUsers);
-    mapAdditionalFields(username, answer, options, moderator);
+    await mapAdditionalFields(request, answer, options);
 
     if (events) {
       events.publish({
@@ -139,9 +150,6 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
     }
 
     const username = await getUsername(request, options);
-    const moderator = await isModerator(request, options);
-    const globalEdit =
-      options.config.getOptionalBoolean('qeta.allowGlobalEdits') ?? false;
 
     const questionId = Number.parseInt(request.params.id, 10);
     const answerId = Number.parseInt(request.params.answerId, 10);
@@ -149,6 +157,18 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
       response.status(400).send({ errors: 'Invalid id', type: 'body' });
       return;
     }
+    await authorize(
+      request,
+      qetaReadQuestionPermission,
+      options,
+      request.params.id,
+    );
+    await authorize(
+      request,
+      qetaEditAnswerPermission,
+      options,
+      request.params.answerId,
+    );
 
     // Act
     const answer = await database.updateAnswer(
@@ -157,7 +177,6 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
       answerId,
       request.body.answer,
       request.body.images,
-      moderator || globalEdit,
     );
 
     if (!answer) {
@@ -165,7 +184,7 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
       return;
     }
 
-    mapAdditionalFields(username, answer, options, moderator);
+    await mapAdditionalFields(request, answer, options);
 
     // Response
     response.status(201);
@@ -191,9 +210,21 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
         response.status(400).send({ errors: 'Invalid id', type: 'body' });
         return;
       }
+      await authorize(
+        request,
+        qetaReadQuestionPermission,
+        options,
+        request.params.id,
+      );
+      await authorize(
+        request,
+        qetaReadAnswerPermission,
+        options,
+        request.params.answerId,
+      );
+      await authorize(request, qetaCreateCommentPermission, options);
 
       const username = await getUsername(request, options);
-      const moderator = await isModerator(request, options);
       const created = await getCreated(request, options);
       // Act
       const question = await database.getQuestion(username, questionId, false);
@@ -225,7 +256,7 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
         request.body.content,
         followingUsers,
       );
-      mapAdditionalFields(username, answer, options, moderator);
+      await mapAdditionalFields(request, answer, options);
 
       if (events) {
         events.publish({
@@ -252,7 +283,6 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
     async (request, response) => {
       // Validation
       const username = await getUsername(request, options);
-      const moderator = await isModerator(request, options);
       const questionId = Number.parseInt(request.params.id, 10);
       const answerId = Number.parseInt(request.params.answerId, 10);
       const commentId = Number.parseInt(request.params.commentId, 10);
@@ -264,13 +294,30 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
         response.status(400).send({ errors: 'Invalid id', type: 'body' });
         return;
       }
+      await authorize(
+        request,
+        qetaReadQuestionPermission,
+        options,
+        request.params.id,
+      );
+      await authorize(
+        request,
+        qetaReadAnswerPermission,
+        options,
+        request.params.answerId,
+      );
+      await authorize(
+        request,
+        qetaDeleteCommentPermission,
+        options,
+        request.params.commentId,
+      );
 
       // Act
       const answer = await database.deleteAnswerComment(
         answerId,
         commentId,
         username,
-        moderator,
       );
 
       if (!answer) {
@@ -278,7 +325,7 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
         return;
       }
 
-      mapAdditionalFields(username, answer, options, moderator);
+      await mapAdditionalFields(request, answer, options);
 
       // Response
       response.status(201);
@@ -291,14 +338,24 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
     // Validation
     // Act
     const username = await getUsername(request, options);
-    const moderator = await isModerator(request, options);
-    await checkPermissions(request, qetaReadPermission, options);
     const questionId = Number.parseInt(request.params.id, 10);
     const answerId = Number.parseInt(request.params.answerId, 10);
     if (Number.isNaN(questionId) || Number.isNaN(answerId)) {
       response.status(400).send({ errors: 'Invalid id', type: 'body' });
       return;
     }
+    await authorize(
+      request,
+      qetaReadQuestionPermission,
+      options,
+      request.params.id,
+    );
+    await authorize(
+      request,
+      qetaEditAnswerPermission,
+      options,
+      request.params.answerId,
+    );
 
     const answer = await database.getAnswer(answerId, username);
 
@@ -307,7 +364,7 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
       return;
     }
 
-    mapAdditionalFields(username, answer, options, moderator);
+    await mapAdditionalFields(request, answer, options);
 
     // Response
     response.json(answer);
@@ -318,7 +375,6 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
     '/questions/:id/answers/:answerId',
     async (request, response) => {
       // Validation
-      const moderator = await isModerator(request, options);
       const username = await getUsername(request, options);
       const questionId = Number.parseInt(request.params.id, 10);
       const answerId = Number.parseInt(request.params.answerId, 10);
@@ -326,6 +382,19 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
         response.status(400).send({ errors: 'Invalid id', type: 'body' });
         return;
       }
+
+      await authorize(
+        request,
+        qetaReadQuestionPermission,
+        options,
+        request.params.id,
+      );
+      await authorize(
+        request,
+        qetaDeleteAnswerPermission,
+        options,
+        request.params.answerId,
+      );
 
       if (events) {
         const question = await database.getQuestion(
@@ -347,11 +416,7 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
       }
 
       // Act
-      const deleted = await database.deleteAnswer(
-        username,
-        answerId,
-        moderator,
-      );
+      const deleted = await database.deleteAnswer(answerId);
 
       // Response
       response.sendStatus(deleted ? 200 : 404);
@@ -370,20 +435,22 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
       response.status(400).send({ errors: 'Invalid id', type: 'body' });
       return;
     }
+    await authorize(
+      request,
+      qetaReadQuestionPermission,
+      options,
+      request.params.id,
+    );
+    await authorize(
+      request,
+      qetaReadAnswerPermission,
+      options,
+      request.params.answerId,
+    );
 
     // Act
-    const username = await getUsername(request, options, true);
-    const moderator = await isModerator(request, options);
-    let voteScore = score;
-    if (moderator && request.query.score) {
-      voteScore = Number.parseInt(request.query.score as string, 10);
-      if (Number.isNaN(voteScore)) {
-        response.status(400).send({ errors: 'Invalid score', type: 'body' });
-        return;
-      }
-    }
-
-    const voted = await database.voteAnswer(username, answerId, voteScore);
+    const username = await getUsername(request, options);
+    const voted = await database.voteAnswer(username, answerId, score);
 
     if (!voted) {
       response.sendStatus(404);
@@ -392,7 +459,7 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
 
     const answer = await database.getAnswer(answerId, username);
 
-    mapAdditionalFields(username, answer, options, moderator);
+    await mapAdditionalFields(request, answer, options);
 
     if (answer === null) {
       response.sendStatus(404);
@@ -441,10 +508,7 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
   router.get(
     `/questions/:id/answers/:answerId/correct`,
     async (request, response) => {
-      const username = await getUsername(request, options, true);
-      const moderator = await isModerator(request, options);
-      const globalEdit =
-        options.config.getOptionalBoolean('qeta.allowGlobalEdits') ?? false;
+      const username = await getUsername(request, options);
       const questionId = Number.parseInt(request.params.id, 10);
       const answerId = Number.parseInt(request.params.answerId, 10);
       if (Number.isNaN(questionId) || Number.isNaN(answerId)) {
@@ -462,12 +526,20 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
         return;
       }
 
-      const marked = await database.markAnswerCorrect(
-        username,
-        questionId,
-        answerId,
-        moderator || globalEdit,
+      await authorize(
+        request,
+        qetaEditQuestionPermission,
+        options,
+        request.params.id,
       );
+      await authorize(
+        request,
+        qetaReadAnswerPermission,
+        options,
+        request.params.answerId,
+      );
+
+      const marked = await database.markAnswerCorrect(questionId, answerId);
 
       if (!marked) {
         response
@@ -502,10 +574,7 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
   router.get(
     `/questions/:id/answers/:answerId/incorrect`,
     async (request, response) => {
-      const username = await getUsername(request, options, true);
-      const moderator = await isModerator(request, options);
-      const globalEdit =
-        options.config.getOptionalBoolean('qeta.allowGlobalEdits') ?? false;
+      const username = await getUsername(request, options);
       const questionId = Number.parseInt(request.params.id, 10);
       const answerId = Number.parseInt(request.params.answerId, 10);
       if (Number.isNaN(questionId) || Number.isNaN(answerId)) {
@@ -513,12 +582,21 @@ export const answersRoutes = (router: Router, options: RouteOptions) => {
         return;
       }
 
-      const marked = await database.markAnswerIncorrect(
-        username,
-        questionId,
-        answerId,
-        moderator || globalEdit,
+      await authorize(
+        request,
+        qetaEditQuestionPermission,
+        options,
+        request.params.id,
       );
+
+      await authorize(
+        request,
+        qetaReadAnswerPermission,
+        options,
+        request.params.answerId,
+      );
+
+      const marked = await database.markAnswerIncorrect(questionId, answerId);
       if (events) {
         const question = await database.getQuestion(
           username,
