+++ plugins/qeta-backend/src/service/routes/questions.ts
@@ -1,16 +1,19 @@
 import { QuestionsOptions } from '../../database/QetaStore';
 import {
-  checkPermissions,
+  authorize,
   getCreated,
   getUsername,
-  isModerator,
   mapAdditionalFields,
 } from '../util';
 import Ajv from 'ajv';
 import { Request, Router } from 'express';
 import {
+  qetaCreateCommentPermission,
   qetaCreateQuestionPermission,
-  qetaReadPermission,
+  qetaDeleteCommentPermission,
+  qetaDeleteQuestionPermission,
+  qetaEditQuestionPermission,
+  qetaReadQuestionPermission,
 } from '@drodil/backstage-plugin-qeta-common';
 import addFormats from 'ajv-formats';
 import {
@@ -31,7 +34,7 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
   router.get(`/questions`, async (request, response) => {
     // Validation
     const username = await getUsername(request, options, true);
-    await checkPermissions(request, qetaReadPermission, options);
+    // TODO: Get Query from permission decision
     const validateQuery = ajv.compile(QuestionsQuerySchema);
     if (!validateQuery(request.query)) {
       response
@@ -60,7 +63,8 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
   router.get(`/questions/list/:type`, async (request, response) => {
     // Validation
     const username = await getUsername(request, options, true);
-    await checkPermissions(request, qetaReadPermission, options);
+    await authorize(request, qetaReadQuestionPermission, options);
+    // TODO: Get Query from permission decision
     const validateQuery = ajv.compile(QuestionsQuerySchema);
     if (!validateQuery(request.query)) {
       response
@@ -97,8 +101,6 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
     // Validation
     // Act
     const username = await getUsername(request, options);
-    const moderator = await isModerator(request, options);
-    await checkPermissions(request, qetaReadPermission, options);
     const questionId = Number.parseInt(request.params.id, 10);
     if (Number.isNaN(questionId)) {
       response
@@ -106,6 +108,12 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
         .send({ errors: 'Invalid question id', type: 'body' });
       return;
     }
+    await authorize(
+      request,
+      qetaReadQuestionPermission,
+      options,
+      request.params.id,
+    );
 
     const question = await database.getQuestion(
       username,
@@ -117,9 +125,11 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
       return;
     }
 
-    mapAdditionalFields(username, question, options, moderator);
-    question.answers?.map(a =>
-      mapAdditionalFields(username, a, options, moderator),
+    await mapAdditionalFields(request, question, options);
+    await Promise.all(
+      (question.answers ?? []).map(
+        async a => await mapAdditionalFields(request, a, options),
+      ),
     );
 
     signalQuestionStats(signals, question);
@@ -133,9 +143,7 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
     // Validation
     // Act
     const username = await getUsername(request, options);
-    const moderator = await isModerator(request, options);
     const created = await getCreated(request, options);
-    await checkPermissions(request, qetaReadPermission, options);
     const validateRequestBody = ajv.compile(CommentSchema);
     const questionId = Number.parseInt(request.params.id, 10);
     if (Number.isNaN(questionId)) {
@@ -151,6 +159,14 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
         .send({ errors: validateRequestBody.errors, type: 'body' });
       return;
     }
+    await authorize(
+      request,
+      qetaReadQuestionPermission,
+      options,
+      request.params.id,
+    );
+    await authorize(request, qetaCreateCommentPermission, options);
+
     const question = await database.commentQuestion(
       questionId,
       username,
@@ -165,9 +181,11 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
       return;
     }
 
-    mapAdditionalFields(username, question, options, moderator);
-    question.answers?.map(a =>
-      mapAdditionalFields(username, a, options, moderator),
+    await mapAdditionalFields(request, question, options);
+    await Promise.all(
+      (question.answers ?? []).map(
+        async a => await mapAdditionalFields(request, a, options),
+      ),
     );
 
     const followingUsers = await database.getUsersForTags(question.tags);
@@ -201,8 +219,7 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
       // Validation
       // Act
       const username = await getUsername(request, options);
-      const moderator = await isModerator(request, options);
-      await checkPermissions(request, qetaReadPermission, options);
+
       const questionId = Number.parseInt(request.params.id, 10);
       const commentId = Number.parseInt(request.params.commentId, 10);
       if (Number.isNaN(questionId) || Number.isNaN(commentId)) {
@@ -210,11 +227,24 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
         return;
       }
 
+      await authorize(
+        request,
+        qetaDeleteQuestionPermission,
+        options,
+        request.params.id,
+      );
+
+      await authorize(
+        request,
+        qetaDeleteCommentPermission,
+        options,
+        request.params.commentId,
+      );
+
       const question = await database.deleteQuestionComment(
         questionId,
         commentId,
         username,
-        moderator,
       );
 
       if (question === null) {
@@ -224,9 +254,11 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
         return;
       }
 
-      mapAdditionalFields(username, question, options, moderator);
-      question.answers?.map(a =>
-        mapAdditionalFields(username, a, options, moderator),
+      await mapAdditionalFields(request, question, options);
+      await Promise.all(
+        (question.answers ?? []).map(
+          async a => await mapAdditionalFields(request, a, options),
+        ),
       );
 
       // Response
@@ -263,7 +295,6 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
   // POST /questions
   router.post(`/questions`, async (request, response) => {
     // Validation
-    await checkPermissions(request, qetaCreateQuestionPermission, options);
     const validateRequestBody = ajv.compile(PostQuestionSchema);
     if (!validateRequestBody(request.body)) {
       response
@@ -271,6 +302,7 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
         .json({ errors: validateRequestBody.errors, type: 'body' });
       return;
     }
+    await authorize(request, qetaCreateQuestionPermission, options);
 
     const tags = getTags(request);
     const entities = getEntities(request);
@@ -333,12 +365,16 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
       return;
     }
 
+    await authorize(
+      request,
+      qetaEditQuestionPermission,
+      options,
+      request.params.id,
+    );
+
     const tags = getTags(request);
     const entities = getEntities(request);
     const username = await getUsername(request, options);
-    const moderator = await isModerator(request, options);
-    const globalEdit =
-      options.config.getOptionalBoolean('qeta.allowGlobalEdits') ?? false;
 
     // Act
     const question = await database.updateQuestion(
@@ -349,7 +385,6 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
       tags,
       entities,
       request.body.images,
-      moderator || globalEdit,
     );
 
     if (!question) {
@@ -376,7 +411,6 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
   // DELETE /questions/:id
   router.delete('/questions/:id', async (request, response) => {
     // Validation
-    const moderator = await isModerator(request, options);
     const username = await getUsername(request, options);
     const questionId = Number.parseInt(request.params.id, 10);
     if (Number.isNaN(questionId)) {
@@ -385,6 +419,12 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
         .send({ errors: 'Invalid question id', type: 'body' });
       return;
     }
+    await authorize(
+      request,
+      qetaDeleteQuestionPermission,
+      options,
+      request.params.id,
+    );
 
     if (events) {
       const question = database.getQuestion(username, questionId, false);
@@ -399,11 +439,7 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
     }
 
     // Act
-    const deleted = await database.deleteQuestion(
-      username,
-      questionId,
-      moderator,
-    );
+    const deleted = await database.deleteQuestion(questionId);
 
     // Response
     response.sendStatus(deleted ? 200 : 404);
@@ -423,9 +459,15 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
       return;
     }
 
+    await authorize(
+      request,
+      qetaReadQuestionPermission,
+      options,
+      request.params.id,
+    );
+
     // Act
     const username = await getUsername(request, options);
-    const moderator = await isModerator(request, options);
     const voted = await database.voteQuestion(username, questionId, score);
 
     if (!voted) {
@@ -440,7 +482,7 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
       return;
     }
 
-    mapAdditionalFields(username, question, options, moderator);
+    await mapAdditionalFields(request, question, options);
     question.ownVote = score;
 
     if (events) {
@@ -474,7 +516,6 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
   // GET /questions/:id/favorite
   router.get(`/questions/:id/favorite`, async (request, response) => {
     const username = await getUsername(request, options);
-    const moderator = await isModerator(request, options);
     const questionId = Number.parseInt(request.params.id, 10);
     if (Number.isNaN(questionId)) {
       response
@@ -482,6 +523,12 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
         .send({ errors: 'Invalid question id', type: 'body' });
       return;
     }
+    await authorize(
+      request,
+      qetaReadQuestionPermission,
+      options,
+      request.params.id,
+    );
 
     const favorited = await database.favoriteQuestion(username, questionId);
 
@@ -496,7 +543,7 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
       false,
     );
 
-    mapAdditionalFields(username, question, options, moderator);
+    await mapAdditionalFields(request, question, options);
 
     // Response
     response.json(question);
@@ -505,7 +552,6 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
   // GET /questions/:id/unfavorite
   router.get(`/questions/:id/unfavorite`, async (request, response) => {
     const username = await getUsername(request, options);
-    const moderator = await isModerator(request, options);
     const questionId = Number.parseInt(request.params.id, 10);
     if (Number.isNaN(questionId)) {
       response
@@ -513,6 +559,12 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
         .send({ errors: 'Invalid question id', type: 'body' });
       return;
     }
+    await authorize(
+      request,
+      qetaReadQuestionPermission,
+      options,
+      request.params.id,
+    );
 
     const unfavorited = await database.unfavoriteQuestion(username, questionId);
 
@@ -527,7 +579,7 @@ export const questionsRoutes = (router: Router, options: RouteOptions) => {
       false,
     );
 
-    mapAdditionalFields(username, question, options, moderator);
+    await mapAdditionalFields(request, question, options);
 
     // Response
     response.json(question);
